+++ packages/numbers/src/models/v1/enums.ts
@@ -2,7 +2,10 @@ export type { TypeEnum as BadRequestTypeEnum } from './bad-request/bad-request';
 export type {
   EventTypeEnum as CallbackPayloadEventTypeEnum,
   FailureCodeEnum as CallbackPayloadFailureCodeEnum,
-  CallbackPayloadStatusEnum,
+  CallbackPayloadActiveNumberStatusEnum,
+  CallbackPayloadNumberOrderStatusEnum,
+  CallbackPayloadNumberOrderEventTypeEnum,
+  ResourceTypeEnum as CallbackPayloadResourceTypeEnum,
 } from './callback-payload/callback-payload';
 export type {
   CodeEnum as InternalErrorErrorCodeEnum,